Rename "createAndUpdateChannel" flow to "channelsAndVideos"
