+++ tests/network-tests/src/scenarios/cli.ts
@@ -1,7 +1,7 @@
-import channelsAndVideos from '../flows/clis/createAndUpdateChannel'
+import channelsAndVideos from '../flows/clis/channelsAndVideos'
 import { scenario } from '../Scenario'
 
 // eslint-disable-next-line @typescript-eslint/no-floating-promises
 scenario('Cli', async ({ job }) => {
-  job('Manage channels and videos through CLI', channelsAndVideos)
+  job('manage channels and videos through CLI', channelsAndVideos)
 })
