+++ tests/network-tests/src/scenarios/full.ts
@@ -28,7 +28,7 @@ import expireProposal from '../flows/proposals/expireProposal'
 import proposalsDiscussion from '../flows/proposalsDiscussion'
 import initDistributionBucket from '../flows/clis/initDistributionBucket'
 import initStorageBucket from '../flows/clis/initStorageBucket'
-import createChannel from '../flows/clis/createAndUpdateChannel'
+import channelsAndVideos from '../flows/clis/channelsAndVideos'
 import { scenario } from '../Scenario'
 import activeVideoCounters from '../flows/content/activeVideoCounters'
 import nftAuctionAndOffers from '../flows/content/nftAuctionAndOffers'
@@ -97,9 +97,11 @@ scenario('Full', async ({ job, env }) => {
   // Content directory
   // following jobs must be run sequentially due to some QN queries that could interfere
   const videoCategoriesJob = job('video categories', testVideoCategories).requires(sudoHireLead)
-  const createChannelJob = job('create channel via CLI', createChannel).requires(videoCategoriesJob)
-  job('add and update video subtitles', addAndUpdateVideoSubtitles).requires(createChannelJob)
-  const videoCountersJob = job('check active video counters', activeVideoCounters).requires(createChannelJob)
+  const channelsAndVideosCliJob = job('manage channels and videos through CLI', channelsAndVideos).requires(
+    videoCategoriesJob
+  )
+  job('add and update video subtitles', addAndUpdateVideoSubtitles).requires(channelsAndVideosCliJob)
+  const videoCountersJob = job('check active video counters', activeVideoCounters).requires(channelsAndVideosCliJob)
   const nftAuctionAndOffersJob = job('nft auction and offers', nftAuctionAndOffers).after(videoCountersJob)
   const commentsAndReactionsJob = job('video comments and reactions', commentsAndReactions).after(
     nftAuctionAndOffersJob
